refactor(zgsm): extract getZgsmSelectedModelInfo to shared utility
